refactor: share app skeleton loading gate
