+++ src/pages/inbox/sidebar/BaseSidebarScreen.tsx
@@ -4,9 +4,8 @@ import TopBarWithLoadingBar from '@components/Navigation/TopBarWithLoadingBar';
 import OptionsListSkeletonView from '@components/OptionsListSkeletonView';
 import ScreenWrapper from '@components/ScreenWrapper';
 
-import {useIsAppLoadPending} from '@hooks/useInFlightRequests';
+import {useAppLoadSkeletonState} from '@hooks/useInFlightRequests';
 import useLocalize from '@hooks/useLocalize';
-import useOnyx from '@hooks/useOnyx';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
 
@@ -15,8 +14,6 @@ import {getSpan} from '@libs/telemetry/activeSpans';
 import type {SkeletonSpanReasonAttributes} from '@libs/telemetry/useSkeletonSpan';
 
 import CONST from '@src/CONST';
-import ONYXKEYS from '@src/ONYXKEYS';
-import isLoadingOnyxValue from '@src/types/utils/isLoadingOnyxValue';
 
 import React, {useEffect} from 'react';
 import {View} from 'react-native';
@@ -28,13 +25,7 @@ function BaseSidebarScreen() {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
-    const isAppLoadPending = useIsAppLoadPending();
-    const [isLoadingApp = false] = useOnyx(ONYXKEYS.IS_LOADING_APP);
-    const [hasLoadedApp = false, hasLoadedAppMetadata] = useOnyx(ONYXKEYS.HAS_LOADED_APP);
-    const isLoadingHasLoadedApp = isLoadingOnyxValue(hasLoadedAppMetadata);
-    // Keep the request queue as the primary signal. The legacy flag only recovers interrupted cold starts after HAS_LOADED_APP hydrates false.
-    const isColdRestartRecoveryFallback = !hasLoadedApp && isLoadingApp;
-    const shouldShowSkeleton = (!hasLoadedApp && (isAppLoadPending || isLoadingHasLoadedApp)) || isColdRestartRecoveryFallback;
+    const {shouldShowSkeleton, isAppLoadPending, hasLoadedApp, isLoadingHasLoadedApp, isColdRestartRecoveryFallback} = useAppLoadSkeletonState();
 
     // Tag an in-flight inbox-tab navigation span when the app-loading skeleton is shown instead of the
     // report list, so durations that include the openApp wait can be queried separately in Sentry.
